share-recipe-frontend/backend/requirements.txt
aioredis==1.3.1
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
greenlet==3.2.3
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jose==1.0.0
jwt==1.4.0
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pendulum==3.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
